moves test to test folder
